rfctr(html): replace parser
Also consolidate tests
