+++ example-docs/language-docs/eng_spa_mult.eml
@@ -79,7 +79,7 @@ re endowed with reason and conscience and should act towards one another in=
 ty and rights. They are endowed with reason and conscience and should act t=
 owards one another in a spirit of brotherhood. All human beings are born fr=
 ee and equal in dignity and rights. They are endowed with reason and consci=
-ence and should act towards one another in a spirit of brotherhood.<br><br>=
+ence and should act towards one another in a spirit of brotherhood. <p>    =
 All human beings are born free and equal in dignity and rights. They are en=
 dowed with reason and conscience and should act towards one another in a sp=
 irit of brotherhood. &quot;Todos los seres humanos nacen libres e iguales e=
@@ -87,7 +87,7 @@ n dignidad y derechos y, dotados como est=C3=A1n de raz=C3=B3n y conciencia=
 , deben comportarse fraternalmente los unos con los otros. Todos los seres =
 humanos nacen libres e iguales en dignidad y derechos y, dotados como est=
 =C3=A1n de raz=C3=B3n y conciencia, deben comportarse fraternalmente los un=
-os con los otros.&quot;<br><br>All human beings are born free and equal in =
+os con los otros.&quot;</p> <p>All human beings are born free and equal in =
 dignity and rights. They are endowed with reason and conscience and should =
 act towards one another in a spirit of brotherhood. All human beings are bo=
 rn free and equal in dignity and rights. They are endowed with reason and c=
@@ -96,18 +96,18 @@ l human beings are born free and equal in dignity and rights. They are endo=
 wed with reason and conscience and should act towards one another in a spir=
 it of brotherhood. All human beings are born free and equal in dignity and =
 rights. They are endowed with reason and conscience and should act towards =
-one another in a spirit of brotherhood.<br><br>All human beings are born fr=
+one another in a spirit of brotherhood.</p> <p>All human beings are born fr=
 ee and equal in dignity and rights. They are endowed with reason and consci=
 ence and should act towards one another in a spirit of brotherhood. All hum=
 an beings are born free and equal in dignity and rights. They are endowed w=
 ith reason and conscience and should act towards one another in a spirit of=
  brotherhood. All human beings are born free and equal in dignity and right=
 s. They are endowed with reason and conscience and should act towards one a=
-nother in a spirit of brotherhood.<br><br>&quot;Todos los seres humanos nac=
+nother in a spirit of brotherhood.</p> <p>&quot;Todos los seres humanos nac=
 en libres e iguales en dignidad y derechos y, dotados como est=C3=A1n de ra=
 z=C3=B3n y conciencia, deben comportarse fraternalmente los unos con los ot=
 ros. Todos los seres humanos nacen libres e iguales en dignidad y derechos =
 y, dotados como est=C3=A1n de raz=C3=B3n y conciencia, deben comportarse fr=
-aternalmente los unos con los otros.&quot;</div>
+aternalmente los unos con los otros.&quot;</p></div>
 
 --000000000000a0666d0606e4cfb7--
